# backend/openapi.yaml
openapi: "3.0.3"
info:
  title: Duekeep API
  version: 1.0.0
  description: |
    Календарь обязательств. Канон для проверяющего: GET /docs (Swagger UI).
    Спека сверена со спринтами 1–6. Sprint 7: register → admin и свои категории (без seed-items);
    items/categories/audit/dashboard/calendar/notifications — только свой owner_id (в JSON не отдаём).
    Нет org_id, инвайтов и ручек /org.
    Access: Authorization Bearer. Refresh: JSON body и/или cookie duekeep_refresh (body важнее).
servers:
  - url: /
    description: Тот же origin (frontend :80 или backend :8080)
...
paths:
...
  /api/v1/kinds/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    patch:
      tags: [Catalog]
      summary: Частичное обновление типа (administrator)
      operationId: patchKind
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/KindPatch"
      responses:
        "200":
          description: Обновлённый kind
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Kind"
        "403":
          description: Роль не administrator
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Нет такого id
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Slug занят
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "422":
          description: Валидация полей
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      tags: [Catalog]
      summary: Удалить тип (administrator)
      operationId: deleteKind
      responses:
        "204":
          description: Удалён
        "403":
          description: Роль не administrator
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Нет такого id
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Есть items (Sprint 3)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /api/v1/categories/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    patch:
      tags: [Catalog]
      summary: Имя, порядок или родитель (admin)
      operationId: patchCategory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryPatch"
      responses:
        "200":
          description: Обновлённый узел
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
        "403":
          description: Viewer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Нет id или нового родителя
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "422":
          description: Цикл или глубина > 3
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      tags: [Catalog]
      summary: Удалить категорию (admin)
      operationId: deleteCategory
      responses:
        "204":
          description: Удалена
        "403":
          description: Viewer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Нет такого id
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Есть дети или items
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /api/v1/notifications:
    get:
      tags: [Notifications]
      summary: Своя лента уведомлений
      operationId: listNotifications
      parameters:
        - in: query
          name: unread
          schema:
            type: boolean
          description: true — только без read_at
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PerPage"
      responses:
        "200":
          description: Страница
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationList"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/notifications/read-all:
    post:
      tags: [Notifications]
      summary: Пометить свои непрочитанные
      operationId: readAllNotifications
      responses:
        "204":
          description: Готово
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/notifications/{id}/read:
    parameters:
      - $ref: "#/components/parameters/ID"
    post:
      tags: [Notifications]
      summary: Пометить прочитанным
      operationId: readNotification
      responses:
        "204":
          description: Прочитано
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/Unprocessable"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    ID:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PerPage:
      name: per_page
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
...
  responses:
    Unauthorized:
      description: Нет или битый access
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    NotFound:
      description: Нет такого id
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unprocessable:
      description: Валидация
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    AttrField:
      type: object
      required: [key, label, type, required]
      properties:
        key:
          type: string
        label:
          type: string
        type:
          type: string
          enum: [string, number, boolean]
        required:
          type: boolean
    Kind:
      type: object
      required: [id, slug, name, color, attr_schema]
      properties:
        id:
          type: string
          format: uuid
        slug:
          type: string
        name:
          type: string
        color:
          type: string
        attr_schema:
          type: array
          items:
            $ref: "#/components/schemas/AttrField"
...
    KindPatch:
      type: object
      properties:
        slug:
          type: string
        name:
          type: string
        color:
          type: string
        attr_schema:
          type: array
          items:
            $ref: "#/components/schemas/AttrField"
    Category:
      type: object
      required: [id, name, sort_order, children]
      properties:
        id:
          type: string
          format: uuid
        parent_id:
          type: string
          format: uuid
          nullable: true
        name:
          type: string
        sort_order:
          type: integer
        children:
          type: array
          items:
            $ref: "#/components/schemas/Category"
...
    CategoryPatch:
      type: object
      properties:
        name:
          type: string
        sort_order:
          type: integer
        parent_id:
          type: string
          format: uuid
          nullable: true
          description: null — сделать корнем; отсутствие поля — не менять
    APIError:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
          additionalProperties: true
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          $ref: "#/components/schemas/APIError"
...
    Notification:
      type: object
      required: [id, item_id, to_status, title, created_at]
      properties:
        id:
          type: string
          format: uuid
        item_id:
          type: string
          format: uuid
        to_status:
          type: string
          enum: [expiring, expired]
        title:
          type: string
        read_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
    NotificationList:
      type: object
      required: [items, page, per_page, total]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Notification"
        page:
          type: integer
        per_page:
          type: integer
        total:
          type: integer
...
security:
  - bearerAuth: []